Move authz lib out of snapshot restore (#63947)
Created the first entry for the new __packages_do_no_import__
folder that will house ESUIs private packages.
