+++ src/plugins/es_ui_shared/__packages_do_not_import__/authorization/components/with_privileges.tsx
@@ -1,13 +1,25 @@
 /*
- * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
- * or more contributor license agreements. Licensed under the Elastic License;
- * you may not use this file except in compliance with the Elastic License.
+ * Licensed to Elasticsearch B.V. under one or more contributor
+ * license agreements. See the NOTICE file distributed with
+ * this work for additional information regarding copyright
+ * ownership. Elasticsearch B.V. licenses this file to you under
+ * the Apache License, Version 2.0 (the "License"); you may
+ * not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *    http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing,
+ * software distributed under the License is distributed on an
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ * KIND, either express or implied.  See the License for the
+ * specific language governing permissions and limitations
+ * under the License.
  */
 
-import { useContext } from 'react';
+import { MissingPrivileges } from '../types';
 
-import { MissingPrivileges } from '../../../../../common/types';
-import { AuthorizationContext } from './authorization_provider';
+import { useAuthorizationContext } from './authorization_provider';
 
 interface Props {
   /**
@@ -29,7 +41,7 @@ const toArray = (value: string | string[]): string[] =>
   Array.isArray(value) ? (value as string[]) : ([value] as string[]);
 
 export const WithPrivileges = ({ privileges: requiredPrivileges, children }: Props) => {
-  const { isLoading, privileges } = useContext(AuthorizationContext);
+  const { isLoading, privileges } = useAuthorizationContext();
 
   const privilegesToArray: Privilege[] = toArray(requiredPrivileges).map(p => {
     const [section, privilege] = p.split('.');
