+++ src/plugins/es_ui_shared/public/index.ts
@@ -47,6 +47,18 @@ export {
   expandLiteralStrings,
 } from './console_lang';
 
+export {
+  AuthorizationContext,
+  AuthorizationProvider,
+  NotAuthorizedSection,
+  WithPrivileges,
+  Privileges,
+  MissingPrivileges,
+  SectionError,
+  Error,
+  useAuthorizationContext,
+} from './authorization';
+
 /** dummy plugin, we just want esUiShared to have its own bundle */
 export function plugin() {
   return new (class EsUiSharedPlugin {
