+++ x-pack/plugins/snapshot_restore/common/types/index.ts
@@ -8,4 +8,3 @@ export * from './repository';
 export * from './snapshot';
 export * from './restore';
 export * from './policy';
-export * from './privileges';
