+++ x-pack/plugins/snapshot_restore/public/application/components/policy_form/steps/step_logistics.tsx
@@ -21,13 +21,13 @@ import {
 } from '@elastic/eui';
 
 import { Repository } from '../../../../../common/types';
-import { CronEditor } from '../../../../shared_imports';
+import { CronEditor, SectionError } from '../../../../shared_imports';
 import { useServices } from '../../../app_context';
 import { DEFAULT_POLICY_SCHEDULE, DEFAULT_POLICY_FREQUENCY } from '../../../constants';
 import { useLoadRepositories } from '../../../services/http';
 import { linkToAddRepository } from '../../../services/navigation';
 import { documentationLinksService } from '../../../services/documentation';
-import { SectionLoading, SectionError } from '../../';
+import { SectionLoading } from '../../';
 import { StepProps } from './';
 
 export const PolicyStepLogistics: React.FunctionComponent<StepProps> = ({
