+++ x-pack/plugins/snapshot_restore/public/application/components/repository_form/type_settings/index.tsx
@@ -6,11 +6,11 @@
 import React from 'react';
 import { FormattedMessage } from '@kbn/i18n/react';
 
-import { REPOSITORY_TYPES } from '../../../../../common/constants';
+import { REPOSITORY_TYPES } from '../../../../../common';
 import { Repository, RepositoryType, EmptyRepository } from '../../../../../common/types';
+import { SectionError } from '../../../../shared_imports';
 import { useServices } from '../../../app_context';
 import { RepositorySettingsValidation } from '../../../services/validation';
-import { SectionError } from '../../index';
 
 import { AzureSettings } from './azure_settings';
 import { FSSettings } from './fs_settings';
