--- x-pack/plugins/snapshot_restore/public/application/lib/authorization/components/index.ts
@@ -1,11 +1,0 @@
-/*
- * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
- * or more contributor license agreements. Licensed under the Elastic License;
- * you may not use this file except in compliance with the Elastic License.
- */
-
-export { AuthorizationProvider, AuthorizationContext } from './authorization_provider';
-
-export { WithPrivileges } from './with_privileges';
-
-export { NotAuthorizedSection } from './not_authorized_section';
