+++ x-pack/plugins/snapshot_restore/public/application/sections/repository_add/repository_add.tsx
@@ -12,7 +12,9 @@ import { RouteComponentProps } from 'react-router-dom';
 import { EuiPageBody, EuiPageContent, EuiSpacer, EuiTitle } from '@elastic/eui';
 import { Repository, EmptyRepository } from '../../../../common/types';
 
-import { RepositoryForm, SectionError } from '../../components';
+import { SectionError } from '../../../shared_imports';
+
+import { RepositoryForm } from '../../components';
 import { BASE_PATH, Section } from '../../constants';
 import { breadcrumbService, docTitleService } from '../../services/navigation';
 import { addRepository } from '../../services/http';
