+++ x-pack/plugins/snapshot_restore/public/application/services/http/use_request.ts
@@ -9,11 +9,10 @@ import {
   UseRequestConfig,
   sendRequest as _sendRequest,
   useRequest as _useRequest,
+  Error as CustomError,
 } from '../../../shared_imports';
 
-import { Error as CustomError } from '../../components/section_error';
-
-import { httpService } from './index';
+import { httpService } from '.';
 
 export const sendRequest = (config: SendRequestConfig) => {
   return _sendRequest<any, CustomError>(httpService.httpClient, config);
