+++ x-pack/plugins/snapshot_restore/public/shared_imports.ts
@@ -12,4 +12,10 @@ export {
   useRequest,
   CronEditor,
   DAY,
+  SectionError,
+  Error,
+  WithPrivileges,
+  useAuthorizationContext,
+  NotAuthorizedSection,
+  AuthorizationProvider,
 } from '../../../../src/plugins/es_ui_shared/public';
